# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Marketplace API",
        "version": "0.0.1",
...
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
...
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
        "/auth/authenticate": {
            "post": {
                "operationId": "auth.authenticate",
                "description": "Authenticate a user using social oauth provider and token.",
                "summary": "Authenticate user via social provider",
                "tags": [
                    "Authentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Requests.AuthenticateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/Resources.UserResource"
                                        },
                                        "authorization": {
                                            "type": "object",
                                            "properties": {
                                                "access_token": {
                                                    "type": "object"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "const": "bearer"
                                                },
                                                "expires_in": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "access_token",
                                                "token_type",
                                                "expires_in"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "user",
                                        "authorization"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "OAuth authentication token validation failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The domain-specific error code.",
                                            "example": "OAUTH_FAILED"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "OauthException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "OAuth authentication failed. Please check your credentials and try again."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Attempting to login or perform actions with a deactivated account.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The domain-specific error code.",
                                            "example": "ACCOUNT_DEACTIVATED"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AccountDeactivatedException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "Your account has been deactivated. Please contact support."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The specific error code identifying a validation failure.",
                                            "example": "VALIDATION_ERROR"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Errors overview message.",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "description": "A detailed description of each field that failed validation.",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "The field is required."
                                                }
                                            }
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "message",
                                        "errors"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "operationId": "auth.refresh",
                "description": "Refresh the user's current authentication token, invalidating the old one and returning a new JWT.",
                "summary": "Refresh token",
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/Resources.UserResource"
                                        },
                                        "authorization": {
                                            "type": "object",
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "integer"
                                                }
                                            },
                                            "required": [
                                                "access_token",
                                                "token_type",
                                                "expires_in"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "user",
                                        "authorization"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Requesting protected routes without a valid JWT token, or token is expired/invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error_code": {
                                            "type": "string",
                                            "description": "The global error code.",
                                            "example": "UNAUTHENTICATED"
                                        },
                                        "exception_type": {
                                            "type": "string",
                                            "description": "The exception class name.",
                                            "example": "AuthenticationException"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "A human-readable error message.",
                                            "example": "You are not authenticated. Please provide a valid token."
                                        }
                                    },
                                    "required": [
                                        "error_code",
                                        "exception_type",
                                        "message"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "Requests.AuthenticateRequest": {
                "type": "object",
                "properties": {
                    "provider": {
                        "type": "string",
                        "enum": [
                            "google",
                            "facebook",
                            "apple"
                        ]
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "required": [
                    "provider",
                    "token"
                ],
                "title": "AuthenticateRequest"
            },
...
            "Resources.UserResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "google_id": {
                        "type": "string"
                    },
                    "facebook_id": {
                        "type": "string"
                    },
                    "is_onboarding_completed": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "phone_iso_code": {
                        "type": "string"
                    },
                    "phone_dial_code": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "username",
                    "avatar",
                    "is_onboarding_completed",
                    "phone_number",
                    "phone_iso_code",
                    "phone_dial_code",
                    "role",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "UserResource"
            }
        }
    }
}